Add files via upload
On-balance volume (OBV) is a technical trading momentum indicator that uses volume flow to predict changes in stock price. OBV measures buying and selling pressure as a cumulative indicator that adds volume on up days and subtracts volume on down days. When the security closes higher than the previous close, all of the day’s volume is considered up-volume. When the security closes lower than the previous close, all of the day’s volume is considered down-volume.

How this indicator works
The actual value of the OBV is unimportant; concentrate on its direction.

When both price and OBV are making higher peaks and higher troughs, the upward trend is likely to continue.
When both price and OBV are making lower peaks and lower troughs, the downward trend is likely to continue.
During a trading range, if the OBV is rising, accumulation may be taking place—a warning of an upward breakout.
During a trading range, if the OBV is falling, distribution may be taking place—a warning of a downward breakout.
When price continues to make higher peaks and OBV fails to make higher peaks, the upward trend is likely to stall or fail. This is called a negative divergence.
When price continues to make lower troughs and OBV fails to make lower troughs, the downward trend is likely to stall or fail. This is called a positive divergence.

Calculation
On Balance Volume is calculated by adding the day's volume to a cumulative total when the security's price closes up, and subtracting the day's volume when the security's price closes down.

If today's close is greater than yesterday's close then:
OBV = Yesterday’s OBV + Today’s Volume

If today’s close is less than yesterday’s close then:
OBV = Yesterday’s OBV – Today’s Volume

If today’s close is equal to yesterday’s close then:
OBV = Yesterday’s OBV